backend/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
bcrypt==4.1.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
